Rename routing module from startup.
