+++ src/lib.rs
@@ -1,2 +1,2 @@
 pub mod challenges;
-pub mod startup;
+pub mod routing;
