+++ src/main.rs
@@ -1,7 +1,7 @@
 use actix_web::web::ServiceConfig;
 use shuttle_actix_web::ShuttleActixWeb;
 
-use cch::startup::config;
+use cch::routing::config;
 
 #[shuttle_runtime::main]
 async fn main() -> ShuttleActixWeb<impl FnOnce(&mut ServiceConfig) + Send + Clone + 'static> {
